otk: introduce `ParserState`, fix nested includes, improve errors
This commit adds a new `ParserState` that is part of the `resolve()`
recursion. It is needed to keep (recursive) state about what
state the parsing is in. Currently this is used to keep track
of what include is used so that nested includes work and also
so that errors can how in what file the issue happend.